Tidy up, move ViewModels to library modules
