+++ client-data/build.gradle
@@ -26,6 +26,8 @@ dependencies {
 
   compile 'com.android.support:support-annotations:25.3.1'
 
+  compile "android.arch.lifecycle:extensions:1.0.0-alpha2"
+
   compile 'com.squareup.retrofit2:retrofit:2.3.0'
   compile 'com.squareup.retrofit2:converter-gson:2.3.0'
   compile 'com.jakewharton.retrofit:retrofit2-rxjava2-adapter:1.0.0'
