+++ client-data/src/main/java/com/jraska/github/client/rx/RxLiveData.java
