+++ client-data/src/main/java/com/jraska/github/client/users/UserViewModelModule.java
@@ -1,17 +1,16 @@
-package com.jraska.github.client.ui;
+package com.jraska.github.client.users;
 
 import android.arch.lifecycle.ViewModel;
 import com.jraska.github.client.Navigator;
 import com.jraska.github.client.analytics.EventAnalytics;
 import com.jraska.github.client.rx.AppSchedulers;
-import com.jraska.github.client.users.UsersRepository;
 import dagger.Module;
 import dagger.Provides;
 import dagger.multibindings.ClassKey;
 import dagger.multibindings.IntoMap;
 
 @Module
-public abstract class UserModelModule {
+public abstract class UserViewModelModule {
   @Provides
   @IntoMap
   @ClassKey(UsersViewModel.class)
